Restructure components into modular folders
