+++ src/components/brain/BrainDiagnosticsPanel.js
