+++ src/components/brain/FrequencyAnalysisGraph.js
