+++ src/components/brain/TimeSlicedBrainView.js
